Rename GET invitation endpoint and move to a new controller
